+++ Assets/InfiniteScrollView/Scripts/Samples~/01_Vertical/TestGUI_01.cs
@@ -14,7 +14,7 @@ public class TestGUI_01 : MonoBehaviour
     {
         infiniteScrollView = FindObjectOfType<InfiniteScrollView>();
         // Init cells first
-        await infiniteScrollView.Initialize();
+        await infiniteScrollView.InitializePool();
     }
 
     private void OnGUI()