+++ Assets/InfiniteScrollView/Scripts/Samples~/02_Horizontal/TestGUI_02.cs
@@ -14,7 +14,7 @@ public class TestGUI_02 : MonoBehaviour
     {
         infiniteScrollView = FindObjectOfType<InfiniteScrollView>();
         // Init cells first
-        await infiniteScrollView.Initialize();
+        await infiniteScrollView.InitializePool();
     }
 
     private void OnGUI()