+++ Assets/InfiniteScrollView/Scripts/Samples~/05_ChatRoom/ChatCell.cs
@@ -4,12 +4,16 @@ using UnityEngine.UI;
 
 public class ChatCell : InfiniteCell
 {
+    // Cell template components
     public Text speakerText;
     public Text messageText;
 
     public override void OnUpdate()
     {
+        // Get cell data
         ChatCellData data = (ChatCellData)CellData.data;
+        
+        // Set cell data to cell template
         speakerText.text = data.speaker;
         messageText.text = data.message;
         speakerText.alignment = data.isSelf ? TextAnchor.UpperRight : TextAnchor.UpperLeft;